chore: move ViewBallotPage to round feature
